Corrigindo o tipo da variável valor para float

+++ Fundamentos da linguagem/Dicionários/Criando_dicionario.py
@@ -5,7 +5,7 @@ for i in range(tam):
     produto = input(f" Informe o nome do {i+1}º produto: ")
     estoque[i] = produto
 
-    valor = int(input(f" Informe o valor do produto: "))
+    valor = float(input(f" Informe o valor do produto: "))
     estoque["produto"] = valor
     print("\n")
